# perf: dont parse matchspec in inner loop