--- a/crates/libsolv_rs/src/pool.rs
+++ b/crates/libsolv_rs/src/pool.rs
@@ -147,7 +147,8 @@ impl<'a> Pool<'a> {
 
         pkgs.sort_by(|&p1, &p2| {
             conda_util::compare_candidates(
-                p1,p2,
+                p1,
+                p2,
                 &self.solvables,
                 &self.names_to_ids,
                 &self.packages_by_name,