Reuse the previously parsed MatchSpec to eliminate re-parsing in an inner loop.


|                                    | libsolv | libsolv-rs | diff         |
|------------------------------------|---------|------------|--------------|
| python=3.9                         | 13.19ms | **8.22ms**    | -4.97ms (35%) |
| tensorboard=2.1.1, grpc-cpp=1.39.1 | 1.56s   | **1.24s**      | -0.31s (20%)  |