@wolfv 2023-07-06
The perf is vs #247 so it's _good_ that it didn't change :)