@ruben-arts 2023-07-06
requirement |libsolv | libsolv-rs
-- | -- | --
python=3.9 |8.5717 ms | 4.1912 ms 
tensorboard=2.1.1, grpc-cpp=1.39.1 | 683.25 ms | 801.73 ms