@tdejager 2023-07-06
Results on M1 Pro:
|                                    | libsolv | libsolv-rs |
|------------------------------------|---------|------------|
| python=3.9                         | 7.3953 ms | **4.2977 ms**    | 
| tensorboard=2.1.1, grpc-cpp=1.39.1 | **500.09 ms** | 758.52 ms  |